refactor: 2단계 1차 피드백 반영 - Lottos 로또번호 관리 / 로또 당첨 통계 로직 분리
